# openapi.yaml
openapi: 3.0.0
info:
  title: Podcast Service API
  version: 1.0.0
  description: Service Standard v1 compliant Podcast Service API for managing podcast shows, episodes, and audio uploads.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AudioUpload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        episodeId:
          type: string
          format: uuid
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
          minimum: 1
        mimeType:
          type: string
        url:
          type: string
          format: uri
        uploadedAt:
          type: string
          format: date-time
      required:
        - id
        - episodeId
        - fileName
        - fileSize
        - mimeType
        - url
        - uploadedAt
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required:
        - type
        - title
        - status
paths:
...
  /shows/{show_id}/episodes/{episode_id}/audio:
    parameters:
      - name: show_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: Upload audio
      description: Upload an audio file for an episode
      tags:
        - audio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file to upload
              required:
                - audio
      responses:
        "201":
          description: Audio uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioUpload"
        "404":
          description: Episode not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"

    get:
      summary: Get audio metadata
      description: Get metadata for an episode's audio file
      tags:
        - audio
      responses:
        "200":
          description: Audio metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioUpload"
        "404":
          description: Audio or episode not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...